% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{descricao}
\alias{descricao}
\alias{descricao_sexo}
\alias{descricao_cor}
\alias{descricao_vinculo}
\alias{descricao_instrucao}
\title{Trocar códigos pela descrição}
\usage{
descricao_sexo(x)

descricao_cor(x)

descricao_vinculo(x)

descricao_instrucao(x)
}
\arguments{
\item{x}{Um vetor de códigos}
}
\value{
O mesmo vetor, com suas descrições
}
\description{
Trocar códigos pela descrição
}
