% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ler_pnadc}
\alias{ler_pnadc}
\title{Ler PNAD continua}
\usage{
ler_pnadc(arquivo, periodo)
}
\arguments{
\item{arquivo}{arquivo com microdados}

\item{periodo}{periodo que será lido}
}
\value{
Uma tibble com 19 variaveis com dados do período indicado
}
\description{
Ler PNAD continua
}
