% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ocupacoes}
\alias{ocupacoes}
\title{Códigos das ocupações da PNAD 2018T2.}
\format{Uma tibble com 614 observações e 3 variáveis:
\describe{
  \item{descricao}{Descrição da ocupação}
  \item{nivel}{Nível de agregação da ocupação}
  \item{codigo}{Código CNAE da ocupação}
}}
\source{
\url{ftp://ftp.ibge.gov.br/Trabalho_e_Rendimento/Pesquisa_Nacional_por_Amostra_de_Domicilios_continua/Trimestral/Microdados/Documentacao/Ocupacao_COD.xls}
}
\usage{
ocupacoes
}
\description{
Um conjunto de dados contendo código CNAE, nível e descrição das ocupações
usadas na PNAD 2018T2.
}
\keyword{datasets}
